chore: deep cleanup and path consolidation (v1.7.4-final)
- Migrate from .config/ to config/ directory structure
- Rename schemas for functional clarity and versioning (v1.7/policy-schema.json, etc.)
- Remove unused scripts: fix_agents.py, validate_real_data.rs
- Update Persistence and Schema Generator to use standard config/ paths
- Finalize agent registry with hierarchical priority and trust-based weighting

+++ src/mcp/mod.rs
@@ -64,12 +64,12 @@ pub struct AgentStatusArgs {
     pub task_id: Option<String>,
 }
 
-#[derive(Debug, Serialize, JsonSchema)]
+#[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
 pub struct AgentSummary {
     pub id: String,
     pub name: String,
     pub availability: String,
-    pub priority: u8,
+    pub priority: u16,
     pub cost: u16,
 }
